Separeted tags and social_network. Moved some parts of code to api library. #580
