+++ libraries/api/CMakeLists.txt
@@ -3,7 +3,10 @@ set(CURRENT_TARGET api)
 list(APPEND CURRENT_TARGET_HEADERS
     include/golos/api/extended_account.hpp
     include/golos/api/account_api_object.hpp
-
+    include/golos/api/comment_api_object.hpp
+    include/golos/api/discussion.hpp
+    include/golos/api/vote_state.hpp
+    include/golos/api/account_vote.hpp
 )
 
 list(APPEND CURRENT_TARGET_SOURCES
