+++ plugins/follow/include/golos/plugins/follow/follow_api_object.hpp
@@ -1,13 +1,14 @@
 #ifndef GOLOS_FOLLOW_API_OBJECT_HPP
 #define GOLOS_FOLLOW_API_OBJECT_HPP
 
-#include <golos/plugins/social_network/api_object/comment_api_object.hpp>
+#include <golos/api/comment_api_object.hpp>
 #include <golos/plugins/follow/follow_objects.hpp>
 #include "follow_forward.hpp"
 
 namespace golos {
     namespace plugins {
         namespace follow {
+            using golos::api::comment_api_object;
 
             struct feed_entry {
                 std::string author;
@@ -18,7 +19,7 @@ namespace golos {
             };
 
             struct comment_feed_entry {
-                social_network::comment_api_object comment;
+                comment_api_object comment;
                 std::vector<std::string> reblog_by;
                 time_point_sec reblog_on;
                 uint32_t entry_id = 0;
@@ -33,7 +34,7 @@ namespace golos {
             };
 
             struct comment_blog_entry {
-                social_network::comment_api_object comment;
+                comment_api_object comment;
                 std::string blog;
                 time_point_sec reblog_on;
                 uint32_t entry_id = 0;
