+++ plugins/tags/include/golos/plugins/tags/tag_visitor.hpp
@@ -5,7 +5,7 @@
 #include <golos/chain/account_object.hpp>
 #include <boost/algorithm/string.hpp>
 
-namespace golos { namespace plugins { namespace social_network { namespace tags {
+namespace golos { namespace plugins { namespace tags {
 
     struct operation_visitor {
         operation_visitor(database& db);
@@ -71,4 +71,4 @@ namespace golos { namespace plugins { namespace social_network { namespace tags 
         } /// ignore all other ops
     };
 
-} } } } // golos::plugins::social_network::tags
+} } } // golos::plugins::tags
